refactor(dto): move DTOs e projections para seus domínios
