+++ src/main/java/com/project/miinhareceita/favorite/dto/FavoriteDTO.java
@@ -1,6 +1,7 @@
-package com.project.miinhareceita.dtos;
+package com.project.miinhareceita.favorite.dto;
 
 import com.project.miinhareceita.favorite.domain.Favorite;
+import com.project.miinhareceita.recipe.dto.RecipeFavoriteDTO;
 
 public class FavoriteDTO {
     private RecipeFavoriteDTO recipeFavoriteDTO;
