+++ src/main/java/com/project/miinhareceita/favorite/dto/FavoriteInsertDTO.java
@@ -1,4 +1,4 @@
-package com.project.miinhareceita.dtos;
+package com.project.miinhareceita.favorite.dto;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
