+++ src/main/java/com/project/miinhareceita/ingredient/projection/IngredientProjection.java
@@ -1,4 +1,4 @@
-package com.project.miinhareceita.projections;
+package com.project.miinhareceita.ingredient.projection;
 
 public interface IngredientProjection {
     Long getId();
