+++ src/main/java/com/project/miinhareceita/services/FavoriteService.java
@@ -1,8 +1,8 @@
 package com.project.miinhareceita.services;
 
-import com.project.miinhareceita.dtos.FavoriteDTO;
-import com.project.miinhareceita.dtos.FavoriteInsertDTO;
 import com.project.miinhareceita.favorite.domain.Favorite;
+import com.project.miinhareceita.favorite.dto.FavoriteDTO;
+import com.project.miinhareceita.favorite.dto.FavoriteInsertDTO;
 import com.project.miinhareceita.favorite.repository.FavoriteRepository;
 import com.project.miinhareceita.recipe.domain.Recipe;
 import com.project.miinhareceita.recipe.repository.RecipeRepository;
