+++ src/main/java/com/project/miinhareceita/services/IngredientService.java
@@ -1,9 +1,9 @@
 package com.project.miinhareceita.services;
 
-import com.project.miinhareceita.dtos.IngredientDTO;
 import com.project.miinhareceita.ingredient.domain.Ingredients;
+import com.project.miinhareceita.ingredient.dto.IngredientDTO;
+import com.project.miinhareceita.ingredient.projection.IngredientProjection;
 import com.project.miinhareceita.ingredient.repository.IngredientsRepository;
-import com.project.miinhareceita.projections.IngredientProjection;
 import com.project.miinhareceita.services.exceptions.DatabaseException;
 import com.project.miinhareceita.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
