+++ src/main/java/com/project/miinhareceita/services/UserService.java
@@ -1,10 +1,10 @@
 package com.project.miinhareceita.services;
 
-import com.project.miinhareceita.dtos.UserDTO;
-import com.project.miinhareceita.dtos.UserInsertDTO;
-import com.project.miinhareceita.projections.UserDetailsProjection;
 import com.project.miinhareceita.user.domain.Role;
 import com.project.miinhareceita.user.domain.User;
+import com.project.miinhareceita.user.dto.UserDTO;
+import com.project.miinhareceita.user.dto.UserInsertDTO;
+import com.project.miinhareceita.user.projection.UserDetailsProjection;
 import com.project.miinhareceita.user.repository.RoleRepository;
 import com.project.miinhareceita.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
